Registro de 'lat' y 'lng' con digitos decimales en polygon Zone

+++ database/migrations/2020_01_20_114424_create_paths_table.php
@@ -16,8 +16,8 @@ class CreatePathsTable extends Migration
         Schema::create('paths', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedInteger('id_zone')->nullable(); 
-            $table->float('lat')->nullable();
-            $table->float('lng')->nullable();
+            $table->double('lat', 15,10)->nullable();
+            $table->double('lng', 15,10)->nullable();
             $table->timestamps();
         });
     }
